♻️ refactor: 아카이브 된 파일들의 import 경로 재설정
